Address FastConformer review feedback
