+++ tests/gpu_test_groups.json
@@ -17,7 +17,7 @@
   },
   "audio": {
     "extras": ["audio_cuda12"],
-    "paths": ["tests/stages/audio", "tests/models/asr/test_nemo_asr_gpu.py"]
+    "paths": ["tests/stages/audio", "tests/models/asr/test_nemo_asr.py"]
   },
   "audio_qwen_asr": {
     "extras": ["audio_cuda12", "vllm"],
