+++ tutorials/audio/nemo_fastconformer/pipeline.yaml
@@ -3,7 +3,6 @@
 # Usage (from the Curator repository root):
 #   python nemo_curator/config/run.py \
 #     --config-path ../../tutorials/audio/nemo_fastconformer \
-#     --config-name pipeline \
 #     manifest_path=tests/fixtures/audio/tagging/sample_input.jsonl
 
 defaults:
